generated fun eval env correction
generated report funs now eval expressions in the function eval env (whose parent env is the enclosing env), not a newly created env whose parent env is the calling env. this fixes an issue when a dbc function is referred to with dbc::fun in another R package and the call cannot find other dbc functions.